function-source/requirements.txt
yt-dlp
google-cloud-storage
fake-useragent
functions-framework
google-api-python-client